+++ parseutil.cpp
@@ -78,7 +78,7 @@ QList<HealLocation>* ParseUtil::parseHealLocs(QString text) {
     for (auto line : lines){
         if (line.contains("MAP_GROUP")){
             QList<QString> li = line.replace(" ","").chopped(2).remove('{').split(',');
-            HealLocation hloc = HealLocation(li[1].remove("MAP_NUM(").remove(")"), i, li[2].toInt(), li[3].toInt());
+            HealLocation hloc = HealLocation(li[1].remove("MAP_NUM(").remove(")"), i, li[2].toUShort(), li[3].toUShort());
             parsed->append(hloc);
             i++;
         }
